Moving monitoring related helpers to dedicated file.
