+++ resources/assets/components/ContentfulEntry/ContentfulEntry.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import { report } from '../../helpers';
 import Loader from '../utilities/Loader';
+import { report } from '../../helpers/monitoring';
 import { withoutNulls } from '../../helpers/data';
 import Affirmation from '../Affirmation/Affirmation';
 import SoftEdgeBlock from '../actions/SoftEdgeBlock';
