+++ resources/assets/middleware/api.js
@@ -3,8 +3,8 @@
 import { get } from 'lodash';
 import { RestApiClient } from '@dosomething/gateway';
 
-import { report } from '../helpers';
 import { PHOENIX_URL } from '../constants';
+import { report } from '../helpers/monitoring';
 import { API } from '../constants/action-types';
 import { getUserToken } from '../selectors/user';
 import { isWithinMinutes } from '../helpers/datetime';
